+++ packages/liveblocks-schema/src/checker/index.ts
@@ -13,7 +13,7 @@ import type {
 import { isBuiltInScalar, visit } from "../ast";
 import { assertNever } from "../lib/assert";
 import { didyoumean as dym } from "../lib/didyoumean";
-import type { ErrorReporter } from "../lib/error-reporting";
+import type { ErrorReporter, Suggestion } from "../lib/error-reporting";
 
 function quote(value: string): string {
   return `'${value.replace(/'/g, "\\'")}'`;
@@ -114,8 +114,8 @@ class Context {
       return "???";
     }
 
-    const startLine = this.errorReporter.lineInfo(range[0]).line1;
-    const endLine = this.errorReporter.lineInfo(range[1]).line1;
+    const startLine = this.errorReporter.toPosition(range[0]).line1;
+    const endLine = this.errorReporter.toPosition(range[1]).line1;
     if (startLine === endLine) {
       return `${startLine}`;
     } else {
@@ -135,11 +135,25 @@ class Context {
     return def;
   }
 
-  report(title: string, range?: Range): void {
+  report(title: string, range?: Range, suggestions?: Suggestion[]): void {
     // FIXME(nvie) Don't throw on the first error! Collect a few (max 3?) and then throw as one error.
     // this.errorReporter.printSemanticError(title, description, range);
-    this.errorReporter.throwSemanticError(title, range);
+    this.errorReporter.throwSemanticError(title, range, suggestions);
   }
+}
+
+function formatReplaceSuggestions(
+  suggestions: string[]
+): Suggestion[] | undefined {
+  if (suggestions.length === 0) {
+    return;
+  }
+
+  return suggestions.map((suggestion) => ({
+    type: "replace",
+    message: `Replace with ${quote(suggestion)}`,
+    value: suggestion,
+  }));
 }
 
 function dupes<T>(items: Iterable<T>, keyFn: (item: T) => string): [T, T][] {
@@ -227,9 +241,11 @@ function checkTypeNameExists(
   if (context.registeredTypes.has(node.name)) {
     return;
   }
+  const suggestions = suggestor(node.name);
   context.report(
     didyoumeanify(`Unknown type ${quote(node.name)}`, suggestor(node.name)),
-    node.range
+    node.range,
+    formatReplaceSuggestions(suggestions)
   );
 }
 
@@ -249,24 +265,25 @@ function checkTypeRefTarget(node: TypeRef, context: Context): void {
 function checkTypeNameIsObjectType(node: TypeName, context: Context): void {
   const def = context.registeredTypes.get(node.name);
   if (!def) {
+    const suggestions = context.suggestors.objectTypeName(node.name);
     context.report(
-      didyoumeanify(
-        `Unknown object type ${quote(node.name)}`,
-        context.suggestors.objectTypeName(node.name)
-      ),
-      node.range
+      didyoumeanify(`Unknown object type ${quote(node.name)}`, suggestions),
+      node.range,
+      formatReplaceSuggestions(suggestions)
     );
     return;
   }
 
   // Check that the payload of a LiveObject type is an object type
   if (def._kind !== "ObjectTypeDefinition") {
+    const suggestions = context.suggestors.objectTypeName(node.name);
     context.report(
       didyoumeanify(
         `Type ${quote(node.name)} is not an object type`,
-        context.suggestors.objectTypeName(node.name)
+        suggestions
       ),
-      node.range
+      node.range,
+      formatReplaceSuggestions(suggestions)
     );
     return undefined;
   }